backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rich>=13.0.0
pymongo>=4.6.0
sentence-transformers>=3.0.0
underthesea>=6.8.0
streamlit>=1.30.0
